+++ LookupAnything/Framework/Fields/RecipesForMachineField.cs
@@ -238,7 +238,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields
         /// <param name="minCount">The minimum number needed or produced.</param>
         /// <param name="maxCount">The maximum number needed or produced.</param>
         /// <param name="chance">The chance of the item being produced.</param>
-        private string GetItemDisplayText(string name, int minCount, int maxCount, int chance)
+        private string GetItemDisplayText(string name, int minCount, int maxCount, decimal chance)
         {
             // get name + count
             string text;
